requirements.txt
python-a2a
openai
python-dotenv 